requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.2
blinker==1.8.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
distlib==0.3.6
distro==1.9.0
Django==5.0.6
django-bootstrap4==23.1
django-crispy-forms==2.0
django-filter==23.5
djangorestframework==3.14.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.44.0
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.9
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.3
humanfriendly==10.0
idna==3.4
imageio==2.31.3
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.3
llvmlite==0.40.1
marisa-trie==1.2.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
narwhals==1.8.1
networkx==3.1
nltk==3.8.1
numba==0.57.1
numpy==1.24.4
onnxruntime==1.15.1
onnxruntime-gpu==1.15.1
openai==0.28.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.2.2
paystack==1.5.0
pdfkit==1.0.0
Pillow==9.4.0
platformdirs==2.6.2
pooch==1.7.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==5.28.1
psycopg2==2.9.5
psycopg2-binary==2.9.6
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
PyMatting==1.1.8
pyparsing==3.1.1
pypng==0.20220715.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.35.1
regex==2024.5.15
requests==2.30.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scipy==1.11.2
shellingham==1.5.4
simplejson==3.19.1
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.4.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.3
srsly==2.4.8
streamlit==1.38.0
sympy==1.12
tenacity==8.5.0
thinc==8.2.4
tifffile==2023.8.30
tokenizers==0.19.1
toml==0.10.2
tornado==6.4.1
tqdm==4.66.1
transformers==4.41.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.17.1
virtualenvwrapper-win==1.2.7
wasabi==1.1.3
watchdog==4.0.2
weasel==0.4.1
wkhtmltopdf==0.2
wrapt==1.16.0
yarl==1.11.1
youtube-dl==2021.12.17
